+++ .storybook/main.ts
@@ -1,0 +1,15 @@
+import type { StorybookConfig } from "@storybook/react-vite";
+
+const config: StorybookConfig = {
+  addons: ["@storybook/addon-docs"],
+  docs: {
+    autodocs: "tag",
+  },
+  framework: {
+    name: "@storybook/react-vite",
+    options: {},
+  },
+  stories: ["../stories/**/*.stories.@(ts|tsx)"],
+};
+
+export default config;