+++ .storybook/preview.ts
@@ -1,0 +1,16 @@
+import type { Preview } from "@storybook/react-vite";
+
+import "../src/styles/styles.css";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/iu,
+        date: /Date$/u,
+      },
+    },
+  },
+};
+
+export default preview;