+++ src/components/index.ts
@@ -2,6 +2,7 @@ export type {
   ButtonBaseProps,
   ButtonColor,
   ButtonProps,
+  ButtonSize,
   ButtonVariant,
 } from "./Button";
 export { Button } from "./Button";
@@ -10,8 +11,9 @@ export type {
   DialogActionsProps,
   DialogInitialFocus,
   DialogProps,
+  DialogState,
   DialogSubmitHandler,
 } from "./Dialog";
 export { Dialog, DialogActions } from "./Dialog";
-export type { IconButtonProps } from "./IconButton";
+export type { IconButtonProps, IconButtonSize } from "./IconButton";
 export { IconButton } from "./IconButton";