+++ src/main.ts
@@ -2,13 +2,16 @@ export type {
   ButtonBaseProps,
   ButtonColor,
   ButtonProps,
+  ButtonSize,
   ButtonVariant,
   DialogActionButtonProps,
   DialogActionsProps,
   DialogInitialFocus,
   DialogProps,
+  DialogState,
   DialogSubmitHandler,
   IconButtonProps,
+  IconButtonSize,
 } from "./components";
 export { Button, Dialog, DialogActions, IconButton } from "./components";
 export type { UseDialogReturn } from "./hooks";